requirements.txt
fastapi
uvicorn[standard]
earthengine-api
google-auth
